+++ __tests__/utils.js
@@ -1,0 +1,5 @@
+const knex = require('../server/infra/database/knex');
+
+module.exports = {
+    knex
+};