+++ database/migrations/sqls/20221218053351-tree-last-name-nullable-up.sql
@@ -1,0 +1,1 @@
+ALTER TABLE tree_denormalized ALTER COLUMN planter_last_name DROP NOT NULL