+++ server/repositories/CaptureRepository.js
@@ -50,6 +50,8 @@ class CaptureRepository extends BaseRepository {
   }
 
   async getStatistics(filter, options = { limit: 3, offset: 0 }) {
+    const knex = this._session.getDB();
+
     const whereBuilder = function (object, builder) {
       const result = builder;
       const filterObject = { ...object };
@@ -70,9 +72,55 @@ class CaptureRepository extends BaseRepository {
         );
         delete filterObject.capture_created_end_date;
       }
+      if (filterObject.planting_organization_uuid) {
+        const stakeholderRelationshipQuery = knex.raw(
+          `SELECT sg.stakeholder_id from stakeholder.getStakeholderChildren(?) sg 
+           join stakeholder.stakeholder ss on ss.id = sg.stakeholder_id 
+           where ss.type = 'Organization'`,
+          [filterObject.planting_organization_uuid],
+        );
+        result.whereIn(
+          'planting_organization_uuid',
+          stakeholderRelationshipQuery,
+        );
+        delete filterObject.planting_organization_uuid;
+      }
       result.where(filterObject);
     };
-    const knex = this._session.getDB();
+
+    const cachmentWhereBuilder = function (object, builder) {
+      const result = builder;
+      const filterObject = { ...object };
+      delete filterObject.card_title;
+      if (filterObject.capture_created_start_date) {
+        result.where(
+          'capture_created_at',
+          '>=',
+          filterObject.capture_created_start_date,
+        );
+        delete filterObject.capture_created_start_date;
+      }
+      if (filterObject.capture_created_end_date) {
+        result.where(
+          'capture_created_at',
+          '<=',
+          filterObject.capture_created_end_date,
+        );
+        delete filterObject.capture_created_end_date;
+      }
+      if (filterObject.planting_organization_uuid) {
+        const regionRelationshipQuery = knex.raw(
+          `SELECT r.name from regions.region r where owner_id in 
+            (SELECT sg.stakeholder_id from stakeholder.getStakeholderChildren(?) sg 
+            join stakeholder.stakeholder ss on ss.id = sg.stakeholder_id 
+            where ss.type = 'Organization')`,
+          [filterObject.planting_organization_uuid],
+        );
+        result.whereIn('catchment', regionRelationshipQuery);
+        delete filterObject.planting_organization_uuid;
+      }
+      result.where(filterObject);
+    };
 
     // total number of growers
     const totalGrowersQuery = knex(this._tableName)
@@ -240,14 +288,14 @@ class CaptureRepository extends BaseRepository {
       .from(function () {
         this.count('* as totalCatchment')
           .from('capture_denormalized')
-          .where((builder) => whereBuilder(filter, builder))
+          .where((builder) => cachmentWhereBuilder(filter, builder))
           .groupBy('catchment')
           .as('catchments');
       });
 
     const topCatchmentQuery = knex(this._tableName)
       .select(knex.raw('catchment, count(*) as count'))
-      .where((builder) => whereBuilder(filter, builder))
+      .where((builder) => cachmentWhereBuilder(filter, builder))
       .whereNotNull('catchment')
       .groupBy('catchment')
       .orderBy('count', 'desc')
@@ -267,7 +315,6 @@ class CaptureRepository extends BaseRepository {
           .from('capture_denormalized')
           .as('planters');
       })
-      // .where((builder) => whereBuilder(filter, builder))
       .groupBy('gender')
       .orderBy('count', 'desc');