+++ server/routes.js
@@ -8,6 +8,11 @@ const {
   captureGrowerApprovalRate,
 } = require('./handlers/captureHandler');
 const { handlerWrapper } = require('./utils/utils');
+const {
+  treeGet,
+  treeStatisticsGet,
+  treeStatisticsGetCard
+} = require('./handlers/treeHandler');
 
 router.get('/capture', handlerWrapper(captureGet));
 router.get('/capture/statistics', handlerWrapper(captureStatisticsGet));
@@ -16,5 +21,9 @@ router.get(
   handlerWrapper(captureStatisticsGetCard),
 );
 router.get('/capture/approvalRate', handlerWrapper(captureGrowerApprovalRate));
+
+router.get('/tree', handlerWrapper(treeGet));
+router.get('/tree/statistics', handlerWrapper(treeStatisticsGet));
+router.get('/tree/statistics/card', handlerWrapper(treeStatisticsGetCard));
 
 module.exports = router;